fix(#51): 인증 토큰이 필요한 유저 서비스 분리
