+++ app/src/main/java/com/ssafy/tott/data/datasource/remote/service/UserService.kt
@@ -1,14 +1,11 @@
 package com.ssafy.tott.data.datasource.remote.service
 
-import com.ssafy.tott.data.model.request.BudgetRequest
 import com.ssafy.tott.data.model.request.LoginRequest
 import com.ssafy.tott.data.model.request.SignupRequest
 import com.ssafy.tott.data.model.request.VerificationRequest
 import com.ssafy.tott.data.model.response.AuthTokenRemoteResponse
-import com.ssafy.tott.data.model.response.BudgetResponse
 import retrofit2.Response
 import retrofit2.http.Body
-import retrofit2.http.GET
 import retrofit2.http.POST
 
 interface UserService {
@@ -36,14 +33,4 @@ interface UserService {
         refreshToken: String,
         // TODO request 내용 작성 필요
     ): Response<AuthTokenRemoteResponse>
-
-    @POST("/api/budget/auth/save")
-    suspend fun fetchSaveBudgetMoney(
-        @Body
-        budgetRequest: BudgetRequest
-    ): Response<Unit>
-
-    @GET("/api/budget/auth/find")
-    suspend fun fetchGetBudgetMoney(
-    ): Response<BudgetResponse>
 }
